Shuffle repository structure
Move some low-level functions deeper into subdirs, remove redundancy
